backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-openai>=0.0.5
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
